nass-1329: tweaks for renames
