+++ packages/settings/src/schema/types.ts
@@ -3,7 +3,7 @@ import * as yup from 'yup';
 export interface Setting<T = any> {
   name?: string;
   description?: string;
-  schema: yup.SchemaOf<T>;
+  type: yup.SchemaOf<T>;
   unit?: string;
   defaultValue: T;
 }
