--- packages/settings/src/schema/validation/index.ts
@@ -1,2 +1,0 @@
-export * from './validateSettings';
-export * from '../global-settings-properties/fields';
